requirements.txt
asgiref==3.8.1
dj-database-url
Django
django-ckeditor
django-js-asset==2.2.0
packaging==24.1
phonenumberslite==8.13.40
pillow==10.4.0
sqlparse==0.2.4
typing_extensions==4.12.2
tzdata==2024.1
whitenoise==6.4
python-dotenv==1.0.1
python-decouple==3.7
boto3==1.26.0
django-storages==1.13.2
email-validator
djongo
dnspython
pymongo==3.12.3
pytz

